Allow multiple command programs
Now you can do something like "a=5;b=6;a*b" or "a=5;b=6;a*b;". So, you
can make multiple assignments and then the last expression will be
returned from the main function. Note that semicolons are used to
separate "lines" of the program and the semicolon at the end is
optional.